修改得利商品添加，让它不支持人民币，让得利币存为浮点型

+++ app/manage/model/ProductionModel.php
@@ -23,11 +23,11 @@ class ProductionModel extends BasicModel
     }
 
     public function setSdlpriceAttr($value){
-        return intval($value);
+        return floatval($value);
     }
 
     public function setRmbpriceAttr($value){
-        return intval($value);
+        return floatval($value);
     }
 
 }
